A simple script for extracting the experiment results from the console output
This script is for extracting the lines form the console output only
directly relevant to the experiment results. It achieves that by
filtering the most of status and debug info on the console, keeping
only the results.
It has to be mentioned that the results extracted by this script CANNOT
be directly put into a spreadsheet. That process is still a manual
labor. However, this script can be used later on as a first step for
auto-generation of a result spreadsheet.